Change OSGi Bundle-Version to reflect PTC build.
Append ${version} with ".ptc-SNAPSHOT" to establish a PTC (snapshot)
build.  Note that, while it may seem more natural to simply alter the
version property, the build's "version-info" target forces a version
that's either undesired or not legitimate from OSGi's perspective.
Consequently, it's less hassle to leave it be and to alter the OSGi
Bundle-Version directly.